Simplified tests (#104)
