--- integration-tests/.gitignore
@@ -1,4 +1,0 @@
-target
-Cargo.lock
-venv
-fixtures
