--- integration-tests/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "array_tests"
-version = "0.1.0"
-license = "Apache-2.0"
-authors = ["Jorge C. Leitao <jorgecarleitao@gmail.com", "Apache Arrow <dev@arrow.apache.org>"]
-readme = "README.md"
-edition = "2018"
-
-[lib]
-bench = false
-
-[dependencies]
-parquet = { path = "../", package = "parquet2" }
