--- integration-tests/README.md
@@ -1,4 +1,0 @@
-# DataPage reader and writer
-
-This crate contains an in-memory format used to test the parquet2
-crate in integration against pyarrow.
